refactor(details): extract image viewer into its own workspace overlay
Move the viewer out of details.rs into `image_viewer.rs` and stop using a
gpui_component dialog for it. A dialog layer occludes the whole window, so it
covered the custom title bar and made the window's close/minimize controls
unclickable, and its fixed viewport-fraction sizing overflowed the bottom of
the window.

The viewer is now held in an `ActiveImageViewer` global and mounted by
`Workspace` as an overlay over its own content — which already sits below the
title bar, so the window controls stay reachable. It fills that area with
`size_full` + `overflow_hidden` and `Contain`-fits the image inside a padded
box, so it sizes to the actual content area instead of hardcoded fractions and
can't overflow. Escape / the close button dismiss it; zoom and pan unchanged.
